Change LUT to uint8 to fix the layer image generation bug

+++ src/gcode/lut.py
@@ -8,26 +8,26 @@ def _build_image_lut_interpolated() -> dict[int, dict[int, np.ndarray]]:
     """
     return {
         2: {
-            0: np.array([[255.0, 255.0, 0.0], [204.0, 153.0, 0.0]], dtype=np.float32),
-            1: np.array([[0.0, 0.0, 0.0], [50.0, 153.0, 0.0]], dtype=np.float32),
+            0: np.array([[255, 255, 0], [204, 153, 0]], dtype=np.uint8),
+            1: np.array([[0, 0, 0], [50, 153, 0]], dtype=np.uint8),
         },
         3: {
-            0: np.array([[255.0, 255.0, 0.0], [204.0, 153.0, 0.0]], dtype=np.float32),
-            1: np.array([[101.0, 102.0, 0.0], [204.0, 204.0, 0.0]], dtype=np.float32),
-            2: np.array([[0.0, 0.0, 0.0], [50.0, 153.0, 0.0]], dtype=np.float32),
+            0: np.array([[255, 255, 0], [204, 153, 0]], dtype=np.uint8),
+            1: np.array([[101, 102, 0], [204, 204, 0]], dtype=np.uint8),
+            2: np.array([[0, 0, 0], [50, 153, 0]], dtype=np.float32),
         },
         4: {
-            0: np.array([[255.0, 255.0, 0.0], [204.0, 153.0, 0.0]], dtype=np.float32),
-            1: np.array([[153.0, 204.0, 0.0], [204.0, 51.0, 0.0]], dtype=np.float32),
-            2: np.array([[50.0, 153.0, 0.0], [255.0, 51.0, 0.0]], dtype=np.float32),
-            3: np.array([[0.0, 0.0, 0.0], [50.0, 153.0, 0.0]], dtype=np.float32),
+            0: np.array([[255, 255, 0], [204, 153, 0]], dtype=np.uint8),
+            1: np.array([[153, 204, 0], [204, 51, 0]], dtype=np.uint8),
+            2: np.array([[50, 153, 0], [255, 51, 0]], dtype=np.uint8),
+            3: np.array([[0, 0, 0], [50, 153, 0]], dtype=np.uint8),
         },
         5: {
-            0: np.array([[255.0, 255.0, 0.0], [204.0, 153.0, 0.0]], dtype=np.float32),
-            1: np.array([[204.0, 0.0, 0.0], [204.0, 153.0, 0.0]], dtype=np.float32),
-            2: np.array([[101.0, 102.0, 0.0], [204.0, 204.0, 0.0]], dtype=np.float32),
-            3: np.array([[50.0, 102.0, 0.0], [101.0, 204.0, 0.0]], dtype=np.float32),
-            4: np.array([[0.0, 0.0, 0.0], [50.0, 153.0, 0.0]], dtype=np.float32),
+            0: np.array([[255, 255, 0], [204, 153, 0]], dtype=np.uint8),
+            1: np.array([[204, 0, 0], [204, 153, 0]], dtype=np.uint8),
+            2: np.array([[101, 102, 0], [204, 204, 0]], dtype=np.uint8),
+            3: np.array([[50, 102, 0], [101, 204, 0]], dtype=np.uint8),
+            4: np.array([[0, 0, 0], [50, 153, 0]], dtype=np.uint8),
         }
     }
